backend/requirements.txt
# ── Core Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# ── Database ────────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.3.2
psycopg2-binary==2.9.9

# ── Authentication ──────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
google-auth==2.29.0

# ── AI / LLM ────────────────────────────────────────────────────────────────────
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-community==0.2.16
langchain-core==0.2.38
langchain-postgres==0.0.9
google-generativeai==0.7.2

# ── Caching ─────────────────────────────────────────────────────────────────────
redis[asyncio]==5.0.8
hiredis==2.3.2

# ── HTTP / Async ────────────────────────────────────────────────────────────────
httpx==0.27.0
sse-starlette==2.1.3

# ── Validation ──────────────────────────────────────────────────────────────────
pydantic==2.7.4
pydantic-settings==2.3.4
email-validator==2.2.0

# ── Serialization ───────────────────────────────────────────────────────────────
orjson==3.10.6

# ── Rate Limiting ───────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Monitoring ──────────────────────────────────────────────────────────────────
sentry-sdk[fastapi]==2.13.0
structlog==24.4.0

# ── Task Queue (optional) ───────────────────────────────────────────────────────
celery[redis]==5.4.0

# ── Testing ─────────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8
